Season25/requirements.txt
pandas>=1.0
numpy>=1.26
scikit-learn>=1.5
nfl-data-py>=0.3.3
pyyaml>=6.0.2
python-dotenv>=1.0.1
matplotlib>=3.9
plotly>=5.24
xgboost>=2.1.1
lightgbm>=4.5.0
fastapi>=0.115
uvicorn[standard]>=0.30
streamlit>=1.38
joblib>=1.4.2
